// src/core/generate/type.ts
export interface Path {
  path: string
  slug: string
}

export interface Rendered {
  raw: string
  slug: string
  meta: MetaValidated
  html: string
  text: string
  excerpt: string | null
}

export interface Content {
  raw: any
  meta: Meta
  excerpt?: string
}

export interface Meta {
  tags?: string[]
  categories?: string[]
  title?: string
  created?: Date
  updated?: Date
}

export interface MetaValidated {
  tags: string[]
  categories: string[]
  title: string
  created: string
  wordCount: number
  updated: string | null
}